backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via starlette
click==8.1.6
    # via uvicorn
fastapi==0.100.1
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
psycopg2-binary==2.9.6
    # via -r requirements.in
pydantic==2.1.1
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.4.0
    # via pydantic
pydantic-settings==2.0.2
    # via -r requirements.in
python-dotenv==1.0.0
    # via pydantic-settings
sniffio==1.3.0
    # via anyio
starlette==0.27.0
    # via fastapi
typing-extensions==4.7.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.23.2
    # via -r requirements.in
